+++ os/net/mac/tsch/tsch-types.h
@@ -146,7 +146,7 @@ enum tsch_timeslot_timing_elements {
 typedef rtimer_clock_t tsch_timeslot_timing_ticks[tsch_ts_elements_count];
 
 /** \brief TSCH timeslot timing elements in micro-seconds */
-typedef uint16_t tsch_timeslot_timing_usec[tsch_ts_elements_count];
+typedef uint32_t tsch_timeslot_timing_usec[tsch_ts_elements_count];
 
 /** \brief Stores data about an incoming packet */
 struct input_packet {
